requirements.txt
flask==3.1.1
flask-cors==6.0.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
mediapipe==0.10.5
moviepy==1.0.3
scenedetect==0.6.2
google-generativeai==0.8.3
gunicorn==21.2.0
python-dotenv==1.0.0